{nixos/,}clickshare-csc1: remove (prepare Qt4 removal)
Qt4 is on it's way out, according to

https://github.com/NixOS/nixpkgs/pull/174634

Barco's ClickShare driver/client requires Qt4;
an update isn't in sight anywhere.

To prepare for the removal of Qt4,
the commit at hand removes the
ClickShare package and its NixOS module.
The release notes are appended with a hint about the
removal and some alternatives that might help users
that are still in need of the driver/client functionality.