unity3d: Remove
The package is non-free, starts but requires an account to work with and
crashes upon exit:

```
$ unity-editor
Segmentation fault (core dumped)
```
The last nixpkgs update in 2019 bumped it to a version from 2018, while
upstream continues to update regularly.

`meta.maintainer` is not empty but this package is effectively
unmaintained.

Unless someone invests into this package, it ought to be removed as it
depends on deprecated GTK2.

See #39976.