abseil-cpp: lock in provided vocabulary types
Abseil provides a special header that allows software distributions
to override compile option dependent features. If this file is left
as-is downstream packages will usually fail to link due to undefined
references.

See the upstream documentation for a more comprehensive explanation:
https://github.com/abseil/abseil-cpp/blob/lts_2021_11_02/absl/base/options.h

Up till now we allowed to bulid abseil itself with different C++
Standards to overcome this problem, but that approach didn't fully
solve it:

 * The override is an optional argument, so its easy to overlook.
 * Processes that see multiple variants of the library (becausse
   they get pulled into the closure by different dependencies) are
   subject to undefined behavior.