release-lib.nix: make packagePlatforms respect badPlatforms
packagePlatforms previously ignored badPlatforms, probably because it is
a fairly [recent] addition. While check-meta.nix doesn't implement it
using subtractLists, it is basically equivalent to the following logic:

  platformSet - badPlatformSet (= effectivePlatformSet)

which we implement using subtractLists (which has somewhat confusing
currying-optimized argument order).

This flaw was discovered when testing #177901 which heavily uses
badPlatforms.

[recent]: https://github.com/NixOS/nixpkgs/commit/b0482248fefbf3b6cdd9c92053cfb49778a3a3a8