xanmod-kernels: add stableTT variant
the xanmod-tt kernel used to be the default but was removed in #165257.
this commit adds back the xanmod1-tt stable variant as linux_xanmod_tt,
leaving the default linux_xanmod kernel to the base stable variant. note
that the TT scheduler patches have not been made available for the edge
variant upstream so we can't add a matching linux_xanmod_latest_tt
variant.