nixos/security: enable pam_mount when cryptHomeLuks is set
Automatically enable `pam_mount`, when `cryptHomeLuks` is set for a
user.

Rational: When `users.users.<username>.cryptHomeLuks` is set on its own
without setting `security.pam.mount.enable` to `true`, the given LUKS
device is not mounted on login. But the brief description of the option
`cryptHomeLuks` suggests to the user, its going to be mounted and
mentions nothing of further configuration steps to be taken.