nixos/doc: Make nixos-manual-combined fail easy to inspect
By writing the unchecked outputs before checking them, they will
be written to a store path, which appears in the log, and which
sticks around even if the build fails. Eventually it is GCed, but
until then, you can open the file.

If you run it in a terminal+editor combination like VSCode, the
failure location is just one Ctrl+click away.