autorandr: WIP Fix distutils import warning
Because autorandr still imports distutils, it produces a bunch of
deprecation warnings. we need to add packaging to propagatedBuildInputs,
but because this wasn't wrapped, the wrapper hook did not take care of
adjusting the system path.

The easiest way I could find to fix this was to actually wrap
autorandr. Then I also had to add a patch to disable autorandr
using sys.executable to fork itself in --batch mode. Otherwise it tries to use
python to run the bash wrapper, which doesn't work. This should be
fine in a nixpkgs environment, since nixpkgs sets the shebang explicitly.