lib.types: Add new type `secret`
Secrets are values that are needed as part of a program's
configuration, but should not end up in the world-readable Nix
store. They can for example be passwords or encryption keys. A service
may want to read them either verbatim in its config file or be pointed
to files containing them. The typical way to implement this for
services in NixOS has been to manually implement specific handling for
any secret data the user may need/want to provide, adding specific
options for every possible parameter. This is true even for services
with `settings`-style options (see
https://github.com/NixOS/rfcs/pull/42).

To alleviate this, add a `secret` type that can represent arbitrary
secret values in a config. To create different types of secrets, the
`mkSecret`, `mkEncryptedSecret` and `mkSecretFile` functions are
used. All take a path to a file which should be copied to a
service-readable location and contains secret data. They differ by
what data the file contains or in how it's used:

- `mkSecret` takes a path to a plaintext file which will be
  substituted for its contents during runtime

- `mkEncryptedSecret` takes a path to an encrypted systemd credential
  file which will be substituted for its unencrypted contents during
  runtime

- `mkSecretFile` takes a path to a file which will be kept in path
  form, but updated to reflect where the file is copied during runtime