formats.yaml: Handle secrets
Add a new function to the YAML format, `configWithSecrets`.

`configWithSecrets` automates secret handling for services which use
it. It takes the desired config output path and the Nix value to
export to JSON as inputs, and provides `LoadCredential`-attributes to
use in a systemd service as well as a script that should be run as
part of service start-up. The `LoadCredential`-attributes are used to
import the secret files to the service's credentials folder, so they
can be read without elevated privileges and, in the case of encrypted
credentials, are unencrypted during runtime. The script creates the
config file and performs the secret substitution when run and should
therefore be run as part of a service's ExecStartPre or ExecStart.