maintainers: add a helper script for the options doc conversion
this script can be used to attempt an automatic conversion of option
docs for most modules. it'll also show a diff of options.json before and
after the changes, which should be a good form for checking for unwanted
changes. we specifically show a json diff rather than an xml diff
because newline changes in json are "\n" added are removed, and those
are easier for diff tools to pick out and show in a meaningful way for
this process.

it does *not* check for incorrectly applied changes though, those aren't
easy enough to do automatically for this script.