colord: make the daemon disablable
We can't build the daemon in pkgsMusl, because rustc does not support
dynamic Musl targets[1].  But it still makes sense to support the rest
of colord, because an application should still be able to link against
it to talk to a colord daemon built in some other way (e.g. provided
by a different distro).

[1]: https://github.com/NixOS/nixpkgs/issues/179242