lib/types: Bracket the nested type descriptions for easier reading
Type descriptions were constructed in a way that did not make the
nesting clear. Example:

lazy attribute set of lazy attribute set of string or boolean or
signed integer or list of string or boolean or signed integer or lazy
attribute set of string or boolean or signed integer or list of string
or boolean or signed integer

possible readings:

(lazy attribute set of (lazy attribute set of string)) or boolean or
signed integer or list of string or boolean or signed integer or lazy
attribute set of string or boolean or signed integer or list of string
or boolean or signed integer

lazy attribute set of ((lazy attribute set of (string or boolean)) or
signed integer or list of string or boolean or signed integer or lazy
attribute set of string or boolean or signed integer or list of string
or boolean or signed integer)

Just to name a few.

Now we just put every nested type into parens, which should make it
possible to decipher the output.