restool: also depend on which
The ls-main script uses `which`. It also has /bin/sh as its shebang,
which means bash runs in Bourne-compatible mode and `which` is not a
builtin.

When the ls-* scripts are invoked from a user shell you don't notice
the problem, because /run/current-system/sw/bin/which is on $PATH, but
when they are invoked in a systemd service it would fail:

  .../bin/ls-addni: line 1054: which: command not found