python-dbusmock: refactor check, avoid dep cycle
python-dbusmock relies on bluez for it's checks. Which shouldn't be a
problem, and isn't one normally. However, checkPhase contained a string
substitution that was always constructed regardless of the value of
doCheck (and simply not used if doCheck was false), and so bluez, which
was not supposed to be a dependency, was used as and registered as one
and caused a dependency cycle when trying to add pcsclite as a
dependency of libfido2.

The string substitution has been removed in favor of something that can
remain disabled when doCheck is false.