ipfs: use passthru for repoVersion
The `repoVersion` is only used outside the derivation and not for the build of IPFS itsef. This is exactly the use-case `passthru` was meant for.
It allows updating the `repoVersion` without rebuilding IPFS. This may come in handy if someone forgets to update it and it needs to be updated later.