xf86videointel: Use the `crocus` and `iris` DRI drivers
The i965 driver was removed in Mesa 22, but the xf86videointel driver
hasn't been updated to reflect this. This leads to the following error
when used with the affected hardware:

(EE) AIGLX error: dlopen of /run/opengl-driver/lib/dri/i965_dri.so failed
                  (/run/opengl-driver/lib/dri/i965_dri.so: cannot open
                  shared object file: No such file or directory)
(EE) AIGLX error: unable to load driver i965

To fix this, add a patch which makes the driver return the appropriate
DRI driver name from Mesa 22, i.e. `crocus` for older GPUs and `iris`
for newer.