prosody: use lua 5.2
The lua version was set to 5_2_compat, which is 5.1 with some compat hacks.
As the comment indicates, this was only needed for Prosody 0.10.0,
and we're shipping 0.12.0 now, so it's about time we got rid of this.

In fact, for Prosody 0.12.0, the upstream docs at
https://prosody.im/doc/packagers#lua-versions recommend using lua 5.4.
I tried building that too, but it fails to evaluate because luadbi is marked
as broken in 5.4.