glibcLocales, glibcLocalesUtf8: only define non-null on linux-glibc
Before the change glibcLocales was pulled in on musl (built successfully
but was not needed: musl does not know how to load glibc locales) and
on android (failed build there). Android failed eval due to the
headers -> zip -> libc recursion.

The change limits glibcLocales down to linux && gnu to target linux-glibc.