crosvm.updateScript: stop trying manifest-versions
manifest-versions never seems to contain the release build any more,
so we can't use it to find the version of crosvm being served to CrOS
devices.

Instead, I've changed the update script to take the latest version of
the appropriate crosvm Chrome OS release branch.  This is the branch
that gets served.  Every release, it is branched off from the
"chromeos" branch (which is the one that passes Chrome OS QA), and
then collects any critical fixes over the lifetime of the release.

With this change, I've introduced a new, simplified versioning
scheme, e.g. 100.0.  The tip build is always 1:1 with the Chrome
version, so having both of those is redundant.  The other number is
the number of commits that have been added to the release branch after
branching from the chromeos branch, so that the number will go up if
we update to include a new commit from the same release.