navidrome: Compile server binary and frontend from source
We currently use release tarballs and binaries distributed by Navidrome.
Moving to compilation from source allows e.g. compiling from a different
repo or applying arbitrary patches to Navidrome.

The build process and update script is based on that of gotify-server,
which similarly includes a Go backend and React frontend.