obs-studio-plugins.obs-ndi: use NDI SDK headers directly
obs-ndi bundles the NDI SDK headers to avoid a more complicated build
setup, but this plays against us here because the soversion for the v5
NDI SDK is "5" even though as far as I know it's fully ABI compatible
with the v4 headers in use here.

Work around the v4/v5 incompatibility here by using the upstream SDK
headers instead.