release-cross.nix: add amjoseph as maintainer for powerpc64*, mips64*
The Nixpkgs wiki [says](https://nixos.wiki/wiki/Nixpkgs/Contributing):

> Maintainers will receive emails from Hydra when package builds enters
> a failed state.

I would like to receive those emails if the Hydra builds of the
boostrap-files for mips64el-linux or powerpc64le-linux ever fail to
build.  I figure that adding myself to the maintainers field here is
the easiest way to make that happen.