python3Packages.psycopg: init at 3.0.15
There are two more variants that should be packaged, one that uses
Cython, and one that supports pooled connections. This package is the
pure python version.

Getting the docs to build required some work upstream, which is the
reason for a bunch of the patching we currently carry. I expect this to
go away over time, given that psycopg is very new-ish.