Add curl as dependency to wasm-pack on all systems
On NixOS, wasm-pack currently issues a warning after many commands that
looks like this:

[WARN]: :-) [60] SSL peer certificate or SSH remote key was not OK (SSL certificate problem: unable to get local issuer certificate)

This error is caused by wasm-pack checking for updates in the background
whenever it runs. The `background_check_for_updates` procedure in
wasm-pack uses the curl crate which depends on libcurl, which doesn't
seem to work right on NixOS when the wasm-pack binary isn't linked
against libcurl. Adding curl as a build input removes this warning and
makes the update checking functionality work correctly.