cups-pdf-to-pdf: init at unstable-2021-12-22
Note that cups-pdf refuses to run without root privileges.
To use the binary, one has to either convince cups to
call it with root privileges, or install it suid root.

Also note that currently, this cups-pdf-fork produces
small pdfs with selectable text, as promised.
However, copying the text produces "garbled" text
(characters are randomly reassigned).
This is a known issue and I don't know how to fix it:

https://github.com/alexivkin/CUPS-PDF-to-PDF/issues/7