barman: patch out subprocess wrapper
This patches the command wrapper to prevent it from executing
subprocesses via sys.executable. This is intended to ensure that the
subprocess is using the same Python interpreter as the superprocess.
However, in this case the barman script has been wrapped as a shell
script, which is not executable as Python.

Since our wrapper script already ensures a consistent version of
Python, this internal wrapping is unnecessary, and so we patch it to
execute the command directly.

Fixes #135238