uqm: Switch to building with SDL2
Version 0.8.0 finally has support for SDL2 and thus much better support
for multiple monitors. While I'm currently on my crap^Wlaptop, I can't
test support for it, however when trying UQM with SDL1 it wasn't even
able to create a window:

  Couldn't set OpenGL 800x600x32 video mode: Couldn't find matching GLX visual
  Couldn't set OpenGL 800x600x24 video mode: Couldn't find matching GLX visual
  Couldn't set OpenGL 800x600x16 video mode: Couldn't find matching GLX visual
  Couldn't set any OpenGL 800x600 video mode!
  Could not initialize video: no fallback at start of program!
  Do not know how to display an error box

SDL2 on the other hand works like a charm and it also scales properly
when you resize the window (one thing that has annoyed me with the SDL1
version in the past).

Signed-off-by: aszlig <aszlig@nix.build>