mate: expose all packages in a scope
MATE currently uses a scope, but it doesn't actually expose that scope
to callers, which makes overriding packages in overlays quite difficult
as you have to figure out where each package is plumbed through in the
attrset. You end up with contrived expressions like:

  mate = super.mate // {
    # ...

    extraPackages =
      (self.lib.remove super.mate.whatever super.mate.extraPackages)
      ++ [ self.mate.whatever ];
  };

This change exposes the scope so that we can use overrideScope' in
overlays instead.